Merged recent changes from external project
The big change with this version is the table definitions.  Prior to this
version, within any extension of the dbRecord class, there was a function
called _initialize.  In that function the table structure and relationships
were defined using arrays.  e.g. '_fields', '_tableName', etc.

Now, that function no longer exists, and structure is defined using a static
array instead.  This will allow for static extensions from the parent class
that can refer to the fields without needing to create an actual instance of
itself.  It also simplifies the definition and implements it in a more
intuitive method.