Moved dbTemplate.php into a repository of its own.
It's gotten big enough and elaborate enough that it needed documentation.  This
seems like good enough a time for it to split off as a project of its own.
Further to that, it never fit the rest of weblibs very well, as the rest of the
files therein are all Javascript.

The new repository is /github.com/jacobEwing/dbTemplate