updates on various libs
canvasSprite:  removed.  have not yet added it's newer version in a subdirector
called "canvasprite".

elementSprite:  Just simplified the "trim" function, and wrapped it in an if to
skip definition if it's already there.  Removed the parent level one as it now
belongs in the elementSprite subdirectory.

geometry.js and helperFuncs.js:  Migrated several math/geometry related
functions out of helperFuncs and into geometry.